+++ src/app/common/Header/index.tsx
@@ -1,5 +1,5 @@
 import * as React          from 'react';
-import NavContainer from '../../Nav/containers/NavContainer';
+import NavContainer from '../../nav/containers/NavContainer';
 import HeaderStyles        from './styles';
 
 const Header: React.StatelessComponent<{}> = () => (
