+++ src/app/home/components/NotesList/styles.ts
