+++ src/app/layouts/Base/index.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { Header } from '../../common';
-import ModalContainer from '../../Modal/containers/ModalContainer';
+import ModalContainer from '../../modal/containers/ModalContainer';
 import BaseLayoutStyles from './styles';
 
 const BaseLayout: React.StatelessComponent<{}> = ({children}) => (
