+++ src/app/modal/components/NoteModal/styles.ts
