+++ src/app/modal/interfaces/index.ts
