+++ src/app/modal/redux/reducer.ts
