+++ src/app/modal/redux/types.tsx
