+++ src/app/nav/components/CategoryItem/styles.ts
