+++ src/app/note/components/FullNote/index.tsx
