+++ src/app/note/components/Note/NoteButtons/index.tsx
