+++ src/app/note/components/Note/NoteButtons/styles.ts
