+++ src/app/note/components/Note/index.tsx
