+++ src/app/note/components/Note/styles.ts
