+++ src/app/note/components/index.ts
