+++ src/app/note/interfaces/index.ts
