+++ src/app/note/redux/types.ts
