--- src/firebase.ts
@@ -1,8 +1,0 @@
-import * as firebase from 'firebase';
-
-import { FirebaseConfig } from './config/keys';
-
-firebase.initializeApp(FirebaseConfig);
-
-// const databaseRef = firebase.database().ref();
-// export const todosRef = databaseRef.child("notes");
