+++ src/pages/FullNote.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { withRouter, RouteComponentProps } from 'react-router';
-import NoteContainer from '../app/Note/containers/NoteContainer';
+import NoteContainer from '../app/note/containers/NoteContainer';
 import { BaseLayout } from '../app/layouts';
 import { AppRoute } from '../app/interfaces';
